Fix building conda environment
- Added pytest to list of installed packages
- Fix file extension typo "yml" vs "yaml"
- Make creation of the environment yml specification cross-platform

The last point needs some explanation. When we simply export the environment
without the --from-history option, all packages that are installed in the env will
be listed. This is not cross-platform because many dependencies depend on the specific
system where the conda environment is created. What we most like want is a spec that only
installs the packages we specified in the "conda create" call. Then, we let conda figure
out which additional packages are required on the specific system. Hopefully this works
as I expect it.